git-svn-id: http://svn.habariproject.org/habari/branches/081011-errorhandling/htdocs/system@2741 653ae4dd-d31e-0410-96ef-6bf7bf53c507

+++ classes/acl.php
@@ -208,7 +208,7 @@ class ACL {
 			return $name;
 		}
 		$name = self::normalize_permission( $name );
-		return DB::get_value( 'SELECT id FROM {tokens} WHERE name=?', array( $name ) );
+		return (int) DB::get_value( 'SELECT id FROM {tokens} WHERE name=?', array( $name ) );
 	}
 
 	/**
